--- modelserver/api/__init__.py
@@ -1,3 +1,0 @@
-from .routes import router
-
-__all__ = ["router"]
